+++ packages/map-interface/src/dev/index.ts
@@ -1,26 +1,27 @@
 // Import other components
+import { Switch } from "@blueprintjs/core";
 import hyper from "@macrostrat/hyper";
-import { useMapRef } from "@macrostrat/mapbox-react";
-
-import { setMapPosition } from "@macrostrat/mapbox-utils";
-import { JSONView, useDarkMode } from "@macrostrat/ui-components";
+import {
+  Spacer,
+  useDarkMode,
+  useStoredState
+} from "@macrostrat/ui-components";
 import mapboxgl from "mapbox-gl";
-import { useCallback, useEffect, useState, useMemo } from "react";
-import { LoaderButton } from "~/map-interface/components/navbar";
-import { LocationPanel } from "@macrostrat/map-interface";
-import { FloatingNavbar } from "~/map-interface/components/navbar";
-import { MapAreaContainer } from "~/map-interface/map-page";
-import { PanelCard } from "~/map-interface/map-page/menu";
-import { CoreMapView, MapMarker } from "~/map-interface/map-page/map-view";
+import { useCallback, useState, useEffect } from "react";
+import { buildInspectorStyle, buildXRayStyle } from "./xray";
+import { MapAreaContainer, PanelCard } from "../container";
+import { FloatingNavbar, MapLoadingButton } from "../context-panel";
+import { MapMarker } from "../helpers";
+import { LocationPanel } from "../location-panel";
+import { MapView } from "../map-view";
+import styles from "./main.module.sass";
+import { TileExtentLayer } from "./tile-extent";
 import {
   FeaturePanel,
   FeatureSelectionHandler,
   TileInfo,
 } from "./vector-tile-features";
-import { TileExtentLayer } from "./tile-extent";
 import { MapPosition } from "@macrostrat/mapbox-utils";
-import styles from "../main.module.styl";
-import { ParentRouteButton } from "./utils";
 
 export enum MacrostratVectorTileset {
   Carto = "carto",
@@ -35,82 +36,30 @@ export enum MacrostratRasterTileset {
 
 export const h = hyper.styled(styles);
 
-function initializeMap(args = {}) {
-  const map = new mapboxgl.Map({
-    container: "map",
-    maxZoom: 18,
-    //maxTileCacheSize: 0,
-    logoPosition: "bottom-left",
-    trackResize: true,
-    antialias: true,
-    optimizeForTerrain: true,
-    ...args,
-  });
-
-  //setMapPosition(map, mapPosition);
-  return map;
-}
 
 interface DevMapViewProps {
-  style: mapboxgl.Style;
-  children: React.ReactNode;
-  transformRequest?: mapboxgl.TransformRequestFunction;
-  mapPosition?: MapPosition;
   accessToken?: string;
+  mapPosition?: MapPosition;
 }
 
-export function DevMapView(props: DevMapViewProps): React.ReactElement {
-  const { style, transformRequest, children, accessToken, mapPosition } = props;
-
-  if (accessToken != null) {
-    mapboxgl.accessToken = accessToken;
-  }
-
-  let mapRef = useMapRef();
-
-  //const baseMapURL = getBaseMapStyle(new Set([]), isDarkMode);
-
-  /* HACK: Right now we need this to force a render when the map
-    is done loading
-    */
-  const [mapInitialized, setMapInitialized] = useState(0);
-
-  // Map initialization
-  useEffect(() => {
-    if (style == null) return;
-    mapRef.current = initializeMap({ style, transformRequest });
-    setMapInitialized(mapInitialized + 1);
-  }, [style, transformRequest]);
-
-  // Map style updating
-  useEffect(() => {
-    if (mapRef?.current == null || style == null) return;
-    mapRef?.current?.setStyle(style);
-  }, [mapRef.current, style]);
-
-  useEffect(() => {
-    const map = mapRef.current;
-    if (map == null) return;
-    setMapPosition(map, mapPosition);
-  }, [mapRef.current, mapInitialized]);
-
-  // This seems to do a bit of a poor job at the moment. Maybe because fo caching?
-
-  return h(CoreMapView, null, [children]);
-}
 
 export function DevMapPage({
-  title = null,
+  title = "Map inspector",
   headerElement = null,
   transformRequest = null,
+  mapPosition = null,
+  mapboxToken = null,
+  overlayStyle = null,
   children,
   style,
 }: {
   headerElement?: React.ReactElement;
   transformRequest?: mapboxgl.TransformRequestFunction;
   title?: string;
-  style: mapboxgl.Style;
+  style: mapboxgl.Style | string;
   children?: React.ReactNode;
+  mapboxToken?: string;
+  overlayStyle?: mapboxgl.Style | string;
 }) {
   /* We apply a custom style to the panel container when we are interacting
     with the search bar, so that we can block map interactions until search
@@ -119,16 +68,36 @@ export function DevMapPage({
     the search bar on mobile platforms
   */
 
+  const dark = useDarkMode();
+  const isEnabled = dark?.isEnabled;
+
+  if (mapboxToken != null) {
+    mapboxgl.accessToken = mapboxToken;
+  }
+
+  style ??= isEnabled
+    ? "mapbox://styles/mapbox/dark-v10"
+    : "mapbox://styles/mapbox/light-v10";
+
   const [isOpen, setOpen] = useState(false);
 
-  const [state, setState] = useState({ showTileExtent: false });
-  const { showTileExtent } = state;
+  const [state, setState] = useStoredState("macrostrat:vector-map-inspector", {
+    showTileExtent: false,
+    xRay: false,
+  });
+  const { showTileExtent, xRay } = state;
+
+  const [actualStyle, setActualStyle] = useState(style);
+
+  useEffect(() => {
+    buildInspectorStyle(style, overlayStyle, { mapboxToken, inDarkMode: isEnabled, xRay })
+      .then(setActualStyle);
+  }, [style, xRay, mapboxToken, isEnabled, overlayStyle]);
 
   const [inspectPosition, setInspectPosition] =
     useState<mapboxgl.LngLat | null>(null);
 
   const [data, setData] = useState(null);
-  const isLoading = false; //useAppState((state) => state.core.mapIsLoading);
 
   const onSelectPosition = useCallback((position: mapboxgl.LngLat) => {
     setInspectPosition(position);
@@ -157,8 +126,6 @@ export function DevMapPage({
     );
   }
 
-  const { isEnabled } = useDarkMode();
-
   let tile = null;
   if (showTileExtent && data?.[0] != null) {
     let f = data[0];
@@ -168,20 +135,28 @@ export function DevMapPage({
   return h(
     MapAreaContainer,
     {
-      navbar: h(FloatingNavbar, { className: "searchbar" }, [
-        h([h(ParentRouteButton), headerElement ?? h("h2", title)]),
-        h("div.spacer"),
-        h(LoaderButton, {
+      navbar: h(FloatingNavbar, [
+        headerElement ?? h("h2", title),
+        h(Spacer),
+        h(MapLoadingButton, {
           active: isOpen,
           onClick: () => setOpen(!isOpen),
-          isLoading,
         }),
       ]),
-      contextPanel: h(PanelCard, [children]),
+      contextPanel: h(PanelCard, [
+        h(Switch, {
+          checked: xRay,
+          label: "X-ray mode",
+          onChange() {
+            setState({ ...state, xRay: !xRay });
+          },
+        }),
+        children,
+      ]),
       detailPanel: detailElement,
       contextPanelOpen: isOpen,
     },
-    h(DevMapView, { style, transformRequest }, [
+    h(MapView, { style: actualStyle, transformRequest, mapPosition}, [
       h(FeatureSelectionHandler, {
         selectedLocation: inspectPosition,
         setFeatures: setData,