+++ packages/map-interface/src/dev/vector-tile-features.ts
@@ -1,5 +1,5 @@
 import { Spinner, Switch } from "@blueprintjs/core";
-import { useMapRef } from "@macrostrat/mapbox-react";
+import { useMapRef, useMapStatus } from "@macrostrat/mapbox-react";
 import mapboxgl from "mapbox-gl";
 import hyper from "@macrostrat/hyper";
 import styles from "./main.module.sass";
@@ -39,7 +39,7 @@ export function FeatureSelectionHandler({
   setFeatures: (features: mapboxgl.MapboxGeoJSONFeature[]) => void;
 }) {
   const mapRef = useMapRef();
-  const isLoading = useAppState((state) => state.core.mapIsLoading);
+  const { isLoading } = useMapStatus();
   const prevLocation = usePrevious(selectedLocation);
 
   useEffect(() => {
@@ -144,10 +144,16 @@ function UnitNumber({ value, unit, precision = 1 }) {
   ]);
 }
 
-export function FeaturePanel({ features }) {
+export function FeaturePanel({ features, focusedSource = null, focusedSourceTitle = null }) {
   if (features == null) return null;
-  return h("div.feature-panel", [
-    h(
+
+  let focusedSourcePanel = null;
+  let filteredFeatures = features;
+  let title = "Features";
+
+  if (focusedSource != null) {
+    title = "Basemap features"
+    focusedSourcePanel = h(
       ExpansionPanel,
       {
         title: "Macrostrat features",
@@ -157,16 +163,22 @@ export function FeaturePanel({ features }) {
       [
         h(LoadingAwareFeatureSet, {
           features,
-          sourceID: "burwell",
+          sourceID: focusedSource,
         }),
       ]
-    ),
+    )
+    filteredFeatures = features.filter((d) => d.source != focusedSource);
+  }
+
+
+  return h("div.feature-panel", [
+    focusedSourcePanel,
     h(
       ExpansionPanel,
-      { title: "Basemap features", className: "basemap-features" },
+      { title, className: "basemap-features", expanded: focusedSource == null },
       [
         h(Features, {
-          features: features.filter((d) => d.source != "burwell"),
+          features: filteredFeatures,
         }),
       ]
     ),